nitrogen6x: Add machine definition
This adds the machine definition to Nitrogen6x from Boundary Devices.
More information at:

http://boundarydevices.com/products/nitrogen6x-board-imx6-arm-cortex-a9-sbc/

In order to take advance of same code used for imx6qsabrelite (on
linux-imx, u-boot, imx-firmaware), we use a MACHINEOVERRIDES trick, so
imx6qsabrelite is taken as a fallback for nitrogen6x machine.

To ensure the boot scripts are available at rootfs time, the do_deploy
task of u-boot-script-boundary is added as a dependency of do_rootfs.

Change-Id: I8bd0b80d2ad8e8741391ec178c10abd75218542e
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>