linux-imx (3.0.35): Fix cpu_regulator problem with Wandboard
With a recent merge, the default kernel for mx6dl changed to linux-imx
3.0.35, rev 3.0.0.  In this branch, calls to mx6_cpu_regulator_init() are
required.  This patch adds this call to the Wandboard init file.

Also, fixed some minor things such as adding the header to the Wandboard
board file so that the original authors are credited, and uncommented
the #include for cpu_op-mx6.h which was commented out for some reason.

Signed-off-by: John Weber <rjohnweber@gmail.com>