cfa10036.inc: Fix machine override definition
The order of processing is done from left to right so the
machine-group ought to be used just before the machine itself and the
'=.' operator /prepends/ the string so we need to change this.

The code now uses the SOC_FAMILY to set this as this is how we've been
doing in all other machines which fits in this case.

Change-Id: I049271ae77fc96b97486736f726bf90bfb58c2b6
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>