u-boot-script-boundary: Allow board-specific boot scripts
This update allows board directories to over-ride the boot script
(6x_bootscript-yocto.txt) or upgrade script (6x_upgrade.txt).

This is a bit complicated because the Yocto MACHINE variable
is not the same as the board directory (u-boot/board/boundary/x)
in either the imx6qsabrelite or nitrogen6x-lite case.

The scripts in board/boundary/nitrogen6x/ are used as a default
if a board doesn't override them.

Note that the boot scripts are only partially functional at this
stage, since there are references to named panels which are not
yet supported.

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>