linux-firmware: add bbappend to modify firmware file names
Upstream changes to the brcmfmac driver require a change to the firmware
file names to match the devid and interface type of the Broadcom chipset.
This also makes it easier for one machine to support multiple chipets in
Yocto.

This patch has been changed to adapt to the linux-firmware version
used in Daisy branch which is based on an older version of linux-firmware
repository.

Signed-off-by: John Weber <rjohnweber@gmail.com>
Acked-by: Eric Nelson <eric.nelson@boundarydevices.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>