linux-boundary: Bump SRCREV to 120f379
This update contains a set of back-ports of patches from the
boundary-imx_3.10.17_1.0.2_ga tree as described in commits
d97ad3b and 0e92e1f.

	https://github.com/Freescale/meta-fsl-arm-extra/commit/d97ad3b
	https://github.com/Freescale/meta-fsl-arm-extra/commit/0e92e1f

The first set is the focus of this update, since it contains important
fixes for LCD output. The second contains a long list of updates
to various drivers.

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>