imx6sl-warp: Add machine definition
The WaRP Board is a Wearable Reference Plaform. The board features:

 - Freescale i.MX6 SoloLite processor with 512MB of RAM
 - Freescale FXOS8700CQ 6-axis Xtrinsic sensor
 - Freescale Kinetis KL16 MCU
 - Freescale Xtrinsic MMA955xL intelligent motion sensing platform

The board implements a hybrid architecture to address the evolving
needs of the wearables market. The platform consists of a main board
and an example daughtercard with the ability to add additional
daughtercards for different usage models.

For more information about the project, visit:

 http://www.warpboard.org/

The machine configuration file used to include Yocto Project support to the
board is a consolidation work only.

It only reflects the work done in Linux kernel, U-Boot and FSL Community BSP,
where the hard work is actually done.

Change-Id: I3ef13f7c0959aa3a62b70ca46c7bf79a2cc5b2ed
Signed-off-by: Daiane Angolini <daiane.angolini@freescale.com>
Signed-off-by: Fabio Estevam <fabio.estevam@freescale.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>