colibri-vf: initial machine for Toradex Vybrid modules
The two modules Colibri VF50 and Colibri VF61 are very similar,
with this generic machine called "colibri-vf" both modules are
supported. The bootloader default environment expects the Linux
kernel zImage as well as the device tree files to be located in
the /boot folder of the root file system. Use IMAGE_INSTALL on
the machine level to install them into the root file system by
default.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>