u-boot-boundary/u-boot-script-boundary: update to fcb04d6 (v2014.07)
Update to Boundary Devices U-Boot 2014.07 tree.

This update only contains a single general-purpose change to the
boot script: the addition of an environment variable 'wlmac' which
will be passed to the WL12xx WiFi driver to be used as the mac
address to prevent the need to use the calibrator utility and
manage files that differ across devices.

The U-Boot source files did not change and the version is bumped
only to keep the two recipes in sync and prevent multiple downloads.

If you configure your board like this:
	U-Boot > setenv wlmac 00:de:ad:be:ef:01
	U-Boot > saveenv && boot

You'll get a clause in your kernel command line:
	wlcore.mac=00:de:ad:be:ef:01

All Boundary Devices Nitrogen6x and Nitrogen6_Max boards shipped
with the WL12xx WiFi/BT adapter should have this variable set
before shipment.

If not, please contact Boundary Devices at info@boundarydevices.com

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>