u-boot-congatec: Fix out-of-tree build
Due the possibility of generating several binaries, for U-Boot, the
build needs to be done out-of-tree however the build system of 2013.04
version was not ready for it. Apply a simple fix for the time being.

The out-of-tree mechanism is much more robust in more recent versions
of U-Boot however this is not an option for this set of boards at the
moment.

Change-Id: I3b80e8c07a41f66dd74465dc064bec4c040d853c
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>