imx6sl-warp: Set EXTRA_IMAGEDEPENDS to u-boot
Set EXTRA_IMAGEDEPENDS to "u-boot".

When user copies a .sdcard to the eMMC the bootloader is not updated, but it is
still helpful to generate a U-boot binary and be able to inspect U-boot source code.

The .sdcard image generated after this patch does not include the U-boot binary
which is only built to ease an eventual update.

See how-to override eMMC boot partition on the following link:

   http://git.denx.de/?p=u-boot.git;a=blob;f=board/warp/README

Signed-off-by: Daiane Angolini <daiane.angolini@freescale.com>
Reviewed-by: Fabio Estevam <fabio.estevam@freescale.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>