imx7d-warp7: Add machine definition
The WaRP7 Board is a Wearable Reference Platform. The board features:

 - i.MX 7Solo applications processor (Cortex-A7/Cortex–M4)
 - 8GB eMMC 5.0 and 4Gb LPDDR3
 - WiFi, Bluetooth, BLE, USB-OTG, NFC
 - I/F Camera, MIPI Display, Audio
 - Accelerometer, Barometer, Gyroscope
 - PMIC, Battery charger

For more information about this reference design, please visit:

https://www.element14.com/community/docs/DOC-79058/l/warp-7-the-next-generation-wearable-reference-platform

The machine configuration file used to include Yocto Project support to the
board is a consolidation work only.

It only reflects the work done in Linux kernel, U-Boot and FSL Community BSP,
where the hard work is actually done.

Signed-off-by: Daiane Angolini <daiane.angolini@nxp.com>
Signed-off-by: Fabio Estevam <fabio.estevam@nxp.com>
Signed-off-by: Andres Lemos <andres.lemos@nxp.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>