cfa100xx: Use a force assignment for IMAGE_FSTYPES
The IMAGE_FSTYPES uses a force assignment in the imx-base.inc file so
we ought to override it in the machine file. This ensures we use the
right value even if we include it after imx-base.inc inclusion.

While moving the code around we also drop the explicit ext3 generation
as the it is queued due the filesystem generation dependency chain.

Change-Id: Icb04838b7ca659b056b57a02cc51ffd1657de579
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>