Convert SOC_FAMILY to MACHINEOVERRIDES
The SOC_FAMILY is an indirection which makes the understanding, from
OpenEmbedded versed developers and users, more difficult. Instead of
using the SOC_FAMILY indirection which is rarely used, we now move to
the MACHINEOVERRIDES mechanism as this is in fact what is in use.

This change is done to align with the meta-fsl-arm BSP.

Change-Id: I81fc8b615db6c20a29d266fa68d17c3dd4de3a10
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>