colibri-vf.conf: fix uboot make target
As some variables are now assigned with ?= in colibri-vf.conf and imx-base.inc
make sure that colibri-vf.conf takes precedence.
(e.g. UBOOT_MAKE_TARGET)
UBOOT_MAKE_TARGET is assigned with ?= in imx-base.inc and u-boot.inc. So in
order to have the machine.conf file take precedence use a = assignemnt.
Changing the ?= to ??= in imx-base.inc would always be overridden from
the ?= "all" assignment in u-boot.inc

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
Acked-by: Stefan Agner <stefan.agner@toradex.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>