Revert "imx7s-warp: Does not include u-boot on image"
The WaRP7 board shipped today does not have the boot partition configured
as read only, the contrary of what were expected when I sent this patch.

Because of that, this patch reverts commit 954548d649cd28e1251844eb62632c25bfdc7d4f.

For detail on how to lock/unlock the boot partition:
https://github.com/WaRP7/WaRP7-User-Guide/blob/v1.1/04-Chapter/U-boot.adoc#boot-partition

Signed-off-by: Daiane Angolini <daiane.angolini@nxp.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>