linux-toradex: imx: update to 5.4-2.3.x-imx
Updates to the toradex_5.4-2.3.x-imx kernel branch which is based
on the community supported 5.4-2.3.x-imx which in turn is base on
the NXP downstream kernel from BSP 5.4.24_2.3.0.

Change to use the defconfig which is part of the kernel tree. In order
to provide a defconfig from your bbappend add a file 'defconfig' through
SRC_URI and set 'KBUILD_DEFCONFIG:<machine> = ""'.

Build the device tree with symbols to allow for device tree overlays
which use symbols.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>