machines: follow kernel deployment changes
The kernel binary image is by default deployed into the rootfs, /boot/.
openembedded-core changed the way it is done in a not backwards
compatible way.
I.e. it now RRECOMMENDS instead of RDEPENDS the package which contains
the kernel binary image.
Commit 1c90b27d2c ("kernel: make kernel-base recommend kernel-image, not depend")

So the way to not deploy also changed. Change machines and recipes
which want to prevent kernel deployment accordingly.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>