scripts: avoid pointless LICENSE churn
I was wondering why a bot decided to force-push a commit to a PR of
mine. It turns out the script responsible for generating the LICENSE
file is not deterministic, so depending on random file system layout
we can end up regenerating the LICENSE without any actual change. For
example:

$ diff -u <(git show baf20676~1:LICENSE | sort) <(git show baf20676:LICENSE | sort)

shows that baf20676 didn't provide any change at all in the actual
contents, yet

$ git show --stat baf20676
commit baf20676bcd2b454577449dbf7b0346d7080b880
Author: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Date:   Wed Jan 22 21:42:44 2025 +0000

    Auto-update LICENSE file with current recipe licenses

 LICENSE | 256 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++--------------------------------------------------------------
 1 file changed, 128 insertions(+), 128 deletions(-)

Import fix from meta-freescale
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>